requirements.txt
streamlit
diffusers
transformers
accelerate
scipy
torch
Pillow
invisible_watermark